requirements.txt
sentence-transformers
faiss-cpu
llama-cpp-python
PyMuPDF
streamlit
langchain
